Remove obsolete code that depended on Q3
Clean up most remaining references to Q3, except for the "Q3ScrollView" comments
that are hints as to the contentsMoving(int,int) breakage we have yet to iron
over (it's minor breakage).

Next step, I'll turn off Q3 support and see if I get 50 errors or 50,000.  If
it's not too big, I mean to go right on and dispose of this before the release.
I know we've rooted out and replaced a great heaping lot of Q3 support code
already.  Let's find out what's left!


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11663 cf6a83e7-f017-4305-b0b3-678bde24f230