* BankEditor no long quielty erases bank, percussion, and keymapping names when user changes setting LSB, MSB, and Pecussion boxes.
This would happen when a user interacted with these boxes, exited the editor
then entered the bank editor again.  Now fixed.

* Toggling the pecsussion checkbox actually updates the selected bank correctly

* Some screen refresh issues clean-up as well.

NOTE:
If there are no keyMapping and the user is sitting on a percussion bank,
when the add keyMapping button is pressed, the right side so refresh
and enable the buttons.  It currently does not.

The user should select a different bank then move back to the percussion bank
if Map settings need to be made (click on little keyboard icon and select from
context pop-up menu.)





git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11738 cf6a83e7-f017-4305-b0b3-678bde24f230