Remove the guts from ProgressDialog, leaving the empty skeleton to avoid causing breakage.
After I did some useful audio recordings, and got stuck with a progress dialog
at 99% that had to be canceled, enough was enough.  This is on top of so many
other problems, and it's not possible to sort this out properly in the next 36
hours.

The real code moves to attic/ for now (fate to be determined later), and in the
meantime no progress dialogs are better than stupidly wrong progress dialogs.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11811 cf6a83e7-f017-4305-b0b3-678bde24f230