Patch from Ted Felix.
When you grab a segment by either side and resize it to be either zero-length or
negative length, this results in a zero-length segment which is difficult to
delete and causes other strange problems like repeats stopping for apparently no
reason. This patch fixes this issue by preventing the user from resizing a
segment to any smaller than either the snap size or the smallest note duration.
It also makes it impossible to resize a segment to negative duration. Also
included is some code simplification and comments.
 
There are a number of other zero-length segment problems that I am working on.
Hope to have the patches available as I have time over the next few weeks.
 
Ted.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11901 cf6a83e7-f017-4305-b0b3-678bde24f230