Corrected possible lost note issued for multiple MIDI Note ON messages for a single pitch.
std::map only allows a single value.  So if more than one NOTE ON event for a
pitch is recieved before a NOTE OFF or NOTE ON w/ Velcoity 0 is sent
The last event could be lost in the map

std::multimap takes care of this since it allows duplicates.

Most clean MIDI streams we encounter have Note ON / Note OFF paired, but this
tweek allows for unconventional streams to be handled a little more
gracefully.

NOTE: multimap::find() return the first instance of matching key -- so which
same pitched note it grabs is undefined.  Maybe a comparitor should be used
to suborder the events based on time.  That way longest sounding notes would
turn off first.

I'll leave that to someone else... this works for now.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11906 cf6a83e7-f017-4305-b0b3-678bde24f230