Matrix view: 1st, 3rd, and 5th horitontal indicators respect current segment
When current segment is changed in matrix view, the horitontal lines in the
scene are updates to reflect key / key changes of the current segment.

This helps in identifying start stop time of the current segment.

More adjustments to events may be needed if conflicting keys are present
among the segments in question....hmmm

This is a good start though.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11915 cf6a83e7-f017-4305-b0b3-678bde24f230