* Force change of zero duration note events in toXmlString (which affects saving).
This change now forces all zero duration note events to a duration of 1 during
file save.  This maethod was chosen because changing all of the constructors
and related setters for events is problematic since negative durations are
used in recording, and many event types are indeed zero duration events.

With this and several over related commits, loading and saving of RG files
update note events with a duration of less than 1 (zero in most cases) to a
duration of 1.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11921 cf6a83e7-f017-4305-b0b3-678bde24f230