Using the Split Tool at the beginning or end of a segment will no longer create zero-length segments.
This is the bug that got me involved with rg development.  It's nice to have 
it fixed, but I'm not sure I've done the right thing.  Comments near 
SegmentSplitCommand::isValid() give a brief description of what I think
might be a better approach.  I'd appreciate some guidance here.  Is there
a preferred approach for handling errors within commands?

The test plan for my changes is here:
http://www.tedfelix.com/rosegarden/TESTPLAN



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11941 cf6a83e7-f017-4305-b0b3-678bde24f230