Fix bug #3007112
If a segment was removed from the composition while edited in the event list
editor, the editor tried to change its window title and crashed when trying
to get the track number of the segment in the composition.

Now the editor is closed when the segment is no more in the composition.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11968 cf6a83e7-f017-4305-b0b3-678bde24f230