Fixed rescale MIDI segments from the left bug.
Now events that where hidden do to a resize or moved beyond end of
composition are rescaled and placed in the rescaled segment (but still hidden)

Now rests are also included in the rescale operation.  This allows segments
to retain rests at the begining and ends of segments and also provides a more
consistent GUI user interaction when rescaling these kinds of segments.

NOTE: There could be some notational oddities if the rescaling factor isn't
chosen wisely.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12075 cf6a83e7-f017-4305-b0b3-678bde24f230