Resized segments remember their size even when truncated at composition end.
Rolled back the SegmetnReconfigureCommand with one exception it calls the
new getEndMarkerTime(FALSE).  The culprit was that sometime the actual
end marker was being sent and sometime the composition end marker was sent.

There was no way of knowing which we were getting.  So I put a flag
in the function to bypass end of composition check, then propigated it to all
places that called SegmentReconfigureCommand and sent getEndMarkerTime to use
the FALSE flag.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12077 cf6a83e7-f017-4305-b0b3-678bde24f230