I didn't want to have to resort to this, but a fixed point size of 7 is the lucky number here.  My track numbers no longer have the top third cut off.
I've seen this problem all the way back to KDE, but it seemed to go away with
earlier versions of Qt, and didn't resurface until fairly recently; Qt 4.6 or
4.7.  I have no idea what the real underlying problem is, but it seemed the only
way to get the numbers fully visible was to either limit their size or enlarge
everything else to give them more room.  This was much preferable due to ease of
implementation, and reduced impact on everything else.

If nobody reacts badly to this change, we'll just run with it.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12120 cf6a83e7-f017-4305-b0b3-678bde24f230