Prototype implementation of one possible way of doing linked segments. Test functionality is to do a select mode control-middle-click-drag to make a linked copy of an existing segment, and to do a select mode control-left-click-drag to make an unlinked copy of a linked segment.
New classes:

LinkedSegment - derived from segment, and lives in the composition like a regular segment. It has a QSharedPointer to a LinkedSegmentReference - LinkedSegments which are linked together point to a common shared LinkedSegmentReference. Uses the RefreshStatus mechanism to determine if it has been changed when receiving an updateLinkedSegments() signal (which is emitted just before the commandExecuted() signal). If a LinkedSegment finds it needs refreshing, it updates the LinkedSegmentReference, which in turn updates all the other sibling LinkedSegments. 

LinkedSegmentReference - also derived from segment, not supposed to be a proper segment in the composition though, just a way of coordinating the LinkedSegments. Contains a list of the LinkedSegments linked to it. Gets updates from whichever LinkedSegment changes and passes on the updates to the others. In principle (though not implemented yet) LinkedSegment could have a time squash/stretch or pitch shift relative to the LinkedSegmentReference. The Composition contains a set of QWeakPointers to all LinkedSegmentReference as a convenience for use during serialisation.

SegmentQuickLinkCommand - analogous to SegmentQuickCopyCommand but to generate the LinkedSegments from the Segment being link-copied, or to create another link to an already existing LinkedSegment.

Changed classes of interest: 

Segment - Gave this a virtual copy constructor style clone function, so where copy constructor was used the clone function can be used to do either a deep copy (just in case the seg's actually a LinkedSegment), or a shallow copy (just a Segment with a copy of the events, even if the original's a LinkedSegment).

CompositionView, SegmentTool, SegmentSelector, Command - these contain the hackery to handle the control-middle-mouse-drag to fire the quick link command. As you can see I had to jump through some hoops to cope with the fact that you end up dragging a segment which gets removed from the composition and is replaced by a LinkedSegment. I hope I've not done anything to heinous achieve this - I'm not proud of the dynamic_cast-ing at line 397 of SegmentSelector.cpp by the way.

CommandHistory - contains the emits of updateLinkedSegments() before the emits of commandExecuted(), which causes the synchronisation of the LinkedSegments before their observers are updated as a result of the emit of commandExecuted().

SegmentQuickCopyCommand - this creates a shallow copy of the segment, just to prove that in principle it's possible to "un-link" a segment (although you have to create a copy to do it at the moment).

RefreshStatus - changed default value of m_needsRefresh to false, as it was causing me grief it being true by default. I couldn't see it made any difference to anything else for it to be initialised false, so I changed it.

Changed classes of not much interest:

In many places changed use of Segment copy constructor to use the clone function. Tried to guess when it was appropriate to use the deep or shallow copy option.

Serialisation:

Event, RosegardenDocument and RoseXmlHandler -  made a small effort to get the branch to be able to read linked segments out and in. Did some hackery with xml element names when outputting a LinkedSegmentReference, so that they'd be ignored when read in by previous and trunk versions. I did this because I wanted to check stuff sometimes against the trunk and it was annoying not to be able to load files generated on the branch in the trunk. When loaded into a trunk build, segments which were linked in this branch will just appear as ordinary copies.

git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/linked_segments_ian@12124 cf6a83e7-f017-4305-b0b3-678bde24f230