More work on handling ties in the matrix.
WORKS:
  * If no selection, single clicking on any one in a series of tied notes
    brings all of the notes into the selection automatically.

BROKEN:
  * If existing selection, single clicking more notes does not invoke
    additional tie checking code, because it calls some other as yet
    undiscovered method
  
  * Shift+clicking to add to selection does not invoke tie checking

  * Drag selection does not invoke tie checking

  * Sometimes clicking the right most in a pair of tied notes actually
    CHANGES the length of the first note, and causes it to cease to be
    tied.  Damn strange one this.  I'm not even clicking on the affected
    note, so this is completely baffling so far.

I don't yet have a real plan for addressing any of the above, and I figure this
is going to get even more messy before it's over.  I'll do whatever it takes, no
matter how brutal it is, and we'll worry about making it pretty later.  If
anybody spots any obvious stupidity in my thinking so far though, feel free to
point it out.  I'm totally winging it right now.  Just diving into the heart of
the problem and trying to work out from the middle without a real plan.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12125 cf6a83e7-f017-4305-b0b3-678bde24f230