Single sweep selecitons are working now.
BUGS:

Shift+sweep selections are not, but at a glance seem to be triggering familiar
code that ought to pick up the ties, but apparently is not.

Shift+click and shift+sweep on already selected elements is still broken, and
moreover looks like a real mind bender.  I might just leave this as a loose end.

This code is getting more and more messy, essentially copying the same block of
code to five different places now after adapting it to local conditions where
it's going.  This is going to be a PITA to maintain if we ever want to refactor
the algorithm, as well we might.  Starting at begin() or end() instead of the
entry point we're actually interested in is grossly inefficient, and I bet this
thing bogs like hell on a real segment full of notes.

I guess I should test that, and if it's grotesque, I should probably start
dealing with that now.  Otherwise, I probably intend to just leave it.  I'm not
enjoying this task much.  It makes my head hurt for a relatively obscure bug
we've tolerated for six years.  (Well, six different relatively obscure bugs,
actually.)



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12130 cf6a83e7-f017-4305-b0b3-678bde24f230