Potential implementation of having a linked segment with a transpose relative to the underlying linked reference segment, which can include a relative key change.
New files: SegmentLinkTransposeCommand.{h,cpp} contains commands for transposing a linked segment, and for resetting the transpose on a linked segment.

Added event property so an event can be ignored for linking - used when inserting a clef change in one linked segment only - see ClefInsertionCommand for a use of it

Adding property of Command so result of a command performed on a linked segment can be ignored (other linked segments are not updated with the results). This is used in the link transpose command where the command should only be applied to the link and the sibling links should be left alone. This required passing the command through as a parameter to the update linked segments signal.

Added serialisation in and out for transposition parameters on a linked segment.

Added transposition parameters to linked segment, and handling of transposition for notes and keys.

Added some test UI menus and buttons - in SegmentParameterBox added a collapse frame for transpose change and reset buttons for linked segments. In notation editor, added a menu option for adding a clef change specifically for one linked segment.

git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/linked_segments_ian@12149 cf6a83e7-f017-4305-b0b3-678bde24f230