Fix settings amnesia bug caused by using a hard coded string in several places, instead of using the variable defined in misc/ConfigGroups.h
Let this be a reminder to us all that using hard coded strings as keys is a bad
practice that predisposes us to make stupid mistakes that lead to settings
amnesia bugs like this one.  That's what misc/ConfigGroups.h is for.  Ideally,
it would be even better if we could do the same thing with the hundreds of other
strings that have to match in multiple places, which are used as keys for the
QSettings mechanism.  We have a "makeviable" here and if we're not careful, a
"make_viable" over there, and a bug.

As big of a PITA as it would be, that's actually not a bad idea for a
refactoring project.  We could probably fix several other settings amnesia bugs
prophylactically by pulling every last key into a file, let's call it
misc/ConfigKeys.h, and switching all the hard coded keys over to use those
instead.  I know I've caught myself making this kind of mistake dozens of times
over the years, and I usually catch it in testing while actively working on some
bit of code, but not always.  Once forgotten, these things can be tricky to hunt
down, so I think this ConfigKeys idea would actually be worth a couple of
afternoons for one of you out there who is looking for a little project in order
to feel useful again.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12154 cf6a83e7-f017-4305-b0b3-678bde24f230