A bit inelegant just block copying like that, but it was expedient, and I'm tired.  This handles removing events in the same fashion as adding them, and seems to work quite well.
The only thing I'm noticing that's a bit off is shift+arrow selecting in the
matrix, but I think that's a more general problem.  The selection tends to break
itself when hitting triggers I haven't yet discerned.  I have a comment in there
about the problem from much earlier, so I'm ignoring the issue so far as this
particular work is concerned.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12164 cf6a83e7-f017-4305-b0b3-678bde24f230