Plan of attack, part 1: bypass fine drag completely.
BUGS: If you shift+click+drag a note, it drags out, then plonks back where you
started, leaving an orange shadow where you tried to drag the note to.  This is
from vestigial handling of the Shift modifier, which had two meanings, but now
has only one in this scheme: "Add to an existing selection."

Plus it's entirely too easy to move a note by mistake.  I'll sort that out once
I've cleaned up the playing field and reduced the number of places to look for
the culprit.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12170 cf6a83e7-f017-4305-b0b3-678bde24f230