Regression point.  I'm tired of putting these couts back in every time I back out the latest failed attempt at doing something constructive with this problem.
The whole issue with moving notes by accident seems to be tied in with a
fundamental failure of assumptions on the part of the code.  It's expecting
values where 3 is a meanintful limit, but getting values hundreds of times
higher than that, so that "if blah > 3" may as well mean "if true."  Yet for all
that, adjusting the magic numbers doesn't get anywhere close to the desired
effect.  I can limit the unwanted drags, but it messes up the threshold for
intended drags too, to the point where you have to drag for three bars to effect
a change.  It's schizophrenic.  It's really broken deep down somewhere, but I
haven't fathomed it yet.  Some fundamental assumption is grossly off.

Oh well, nothing else to do on a miserable winter day but keep trying stupid
things and discovering why they didn't work.  Whee!


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12172 cf6a83e7-f017-4305-b0b3-678bde24f230