Phase 3: unraveling a complicated mystery
All that clicky no clicky stuff was a red herring that I probably needn't have
bothered touching at all.  The "too easy to move notes" problem is directly
attributable to StaffLayout::getHeightAtSceneCoords() returning a different
height than the one the clicked element was born with.

It looks like the lazy refinements referred to in comments are a noble thing,
because the "if I returned here" value is always much further off than the one
it's actually returning.  The returned value is quite accurate, but perhaps it's
*too* accurate.

I know what to tinker with now, but don't have a concrete plan.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12174 cf6a83e7-f017-4305-b0b3-678bde24f230