This kludge is a definite improvement.  I compared side by side with an unmodified Rosegarden, and it was considerably easier to move notes by mistake with the older version.  It's still easy to move notes by mistake at smaller staff sizes, and at any size there is a markedly greater tendency to move a note by mistake when clicking toward the top half of the note.
I'm not sure where the note heads get their idea of whether they've been clicked
on or not, but if these limiting rects can be made 10% smaller without creating
chaos, I think that making the hot spot a little smaller will work well
hand-in-hand with this weighted height kludge.

The whole thing really is a complicated mess.  I have a gut feeling that the
fundamental problem with all of this is that QGraphicsScene and friends reckon
everything as QPointF and we're working with ints everywhere for the Y
coordinate.  I think it's probably that we're rounding too early or too many
times, and the end results have lost too much information to make the best
decision what to set the resulting height to.  Playing that hunch looks like a
really expensive proposition, so I'm trying to work with things as they are, and
I finally feel like I made some worthwhile progress at the end of a very tedious
day of gnawing at this absurd little problem.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@12176 cf6a83e7-f017-4305-b0b3-678bde24f230