Changing the underlying mechanism for segment linking, whilst preserving the current functionality.
Removed all classes derived from Segment. Added a SegmentLinker class which handles all the coordination between the linked segments. Segments have a SegmentLinker pointer. If this pointer points to a valid SegmentLinker object then it is linked. If this pointer is null then it is just an ordinary segment. Segments which are linked together point to the same SegmentLinker.

All SegmentLinkers are connected to the signal which is emitted after any command is performed which is the trigger for doing the segment coordination. SegmentLinker uses the refresh status mechanism to determine whether any of its segments have changed, and if so, refreshes all of its other segments from the changed one.

For the time being I've left the transposition stuff in as it was before, in anticipation of it being overhauled in the near future.

The serialisation is simplified from before, all it requires is some extra attributes to be serialised with the segment (a linker id and some transposition parameters). There is no need to serialise anything from the SegmentLinker class. The happy consequence of this is that there will be no problems with forward or backward compatibility. The worst that will happen is that people loading compositions containing linked segments in older versions which don't support linked segments will just see their linked segments appearing as ordinary segments.

git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/linked_segments_ian@12186 cf6a83e7-f017-4305-b0b3-678bde24f230